requirements.txt
jmcomic==2.5.32
Pillow==11.1.0
PyYAML==6.0.2
PyYAML==6.0.2
websockets==15.0.1
